Add ignoring interruption capacity
