--- src/test/java/com/github/junit5docker/InterruptibleRunnable.java
@@ -1,16 +1,0 @@
-package com.github.junit5docker;
-
-@FunctionalInterface
-interface InterruptibleRunnable {
-    void run() throws InterruptedException;
-
-    static Runnable ignoreInterrupted(InterruptibleRunnable callable) {
-        return () -> {
-            try {
-                callable.run();
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            }
-        };
-    }
-}
